requirements.txt
pandas
mysql-connector-python
cassandra-driver